Refactor Usecases
- Merge all uses case into one file. I did not delete individual usecase
classes yet because I need to investigate the suspend method
