--- app/src/main/java/com/faust/m/flashcardm/core/usecase/booklet/ResetForReview.kt
@@ -1,9 +1,0 @@
-package com.faust.m.flashcardm.core.usecase.booklet
-
-import com.faust.m.flashcardm.core.data.CardRepository
-
-class ResetForReview(private val cardRepository: CardRepository) {
-
-    operator fun invoke(count: Int, bookletId: Long) =
-        cardRepository.resetForReview(count, bookletId)
-}
