--- app/src/main/java/com/faust/m/flashcardm/core/usecase/card/UpdateCardContent.kt
@@ -1,9 +1,0 @@
-package com.faust.m.flashcardm.core.usecase.card
-
-import com.faust.m.flashcardm.core.data.CardRepository
-import com.faust.m.flashcardm.core.domain.Card
-
-class UpdateCardContent(private val cardRepository: CardRepository) {
-
-    operator fun invoke(card: Card): Card = cardRepository.updateCardContent(card)
-}
