--- app/src/main/java/com/faust/m/flashcardm/framework/CardUseCases.kt
@@ -1,10 +1,0 @@
-package com.faust.m.flashcardm.framework
-
-import com.faust.m.flashcardm.core.usecase.card.*
-
-class CardUseCases(val addCard: AddCard,
-                   val deleteCard: DeleteCard,
-                   val getCardsForBooklet: GetCardsForBooklet,
-                   val updateCard: UpdateCard,
-                   val updateCardContent: UpdateCardContent
-)
