+++ app/src/main/java/com/faust/m/flashcardm/presentation/fragment_edit_card/ViewModelEditCard.kt
@@ -3,7 +3,7 @@ package com.faust.m.flashcardm.presentation.fragment_edit_card
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import com.faust.m.flashcardm.core.domain.Card
-import com.faust.m.flashcardm.framework.CardUseCases
+import com.faust.m.flashcardm.core.usecase.CardUseCases
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import org.jetbrains.anko.AnkoLogger
