+++ app/src/main/java/com/faust/m/flashcardm/presentation/library/LibraryViewModel.kt
@@ -6,9 +6,9 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Transformations
 import androidx.lifecycle.ViewModel
 import com.faust.m.flashcardm.core.domain.Booklet
-import com.faust.m.flashcardm.core.usecase.booklet.BookletOutline
 import com.faust.m.flashcardm.R
-import com.faust.m.flashcardm.framework.BookletUseCases
+import com.faust.m.flashcardm.core.usecase.BookletOutline
+import com.faust.m.flashcardm.core.usecase.BookletUseCases
 import com.faust.m.flashcardm.framework.FlashViewModel
 import com.faust.m.flashcardm.presentation.Event
 import com.faust.m.flashcardm.presentation.MutableLiveEvent
