+++ app/src/main/java/com/faust/m/flashcardm/presentation/review/ReviewViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.faust.m.flashcardm.core.domain.Card
-import com.faust.m.flashcardm.framework.CardUseCases
+import com.faust.m.flashcardm.core.usecase.CardUseCases
 import com.faust.m.flashcardm.presentation.fragment_edit_card.DelegateEditCard
 import com.faust.m.flashcardm.presentation.fragment_edit_card.ViewModelEditCard
 import com.faust.m.flashcardm.presentation.review.ReviewCard.State.ASKING
