+++ app/src/main/java/com/faust/m/flashcardm/presentation/view_library_booklet/ViewModelLibraryBooklet.kt
@@ -2,8 +2,8 @@ package com.faust.m.flashcardm.presentation.view_library_booklet
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.faust.m.flashcardm.core.usecase.booklet.BookletOutline
-import com.faust.m.flashcardm.framework.BookletUseCases
+import com.faust.m.flashcardm.core.usecase.BookletOutline
+import com.faust.m.flashcardm.core.usecase.BookletUseCases
 import com.faust.m.flashcardm.framework.FlashViewModel
 import com.faust.m.flashcardm.presentation.library.LibraryBooklet
 import kotlinx.coroutines.GlobalScope
